% Generated by roxygen2 (4.0.1): do not edit by hand
\name{colorwheel}
\alias{colorwheel}
\title{Visualize color palettes using a color wheel}
\usage{
colorwheel(chroma = 100, luminance = 60, labels = TRUE, alpha = 1,
  fixup = TRUE)
}
\arguments{
\item{chroma}{chroma value for palette}

\item{luminance}{luminance value for palette}

\item{labels}{logical, should degree labels be shown}

\item{alpha}{transparency level defined on the interval [0, 1] where
0 = transparent and 1 = opaque}

\item{fixup}{logical, should RGB value be corrected (see \code{\link{hcl}}}
}
\value{
\code{colorwheel} returns a pie chart showing the specified colors
}
\description{
\code{colorwheel} allows you to easily visualize a color palette by
    plotting all hues for the specified specified chroma and luminance
    values using a color wheel
}
\examples{
colorwheel()
colorwheel(chroma = 50, luminance = 60)
colorwheel(chroma = 100, luminance = 75)
colorwheel(chroma = 100, luminance = 25)
colorwheel(chroma = 100, luminance = 0)
colorwheel(chroma = 0, luminance = 50)

par(mfrow = c(4, 3), mar = c(2, 0, 1.5, 0))
for(i in seq(0, 100, 10))
   colorwheel(chroma = i, labels = TRUE, fixup = TRUE)

par(mfrow = c(3, 3), mar = c(2, 0, 1.5, 0))
for(i in seq(20, 100, 10))
    colorwheel(luminance = i, labels = TRUE)
}
\seealso{
\code{\link{qpal}} \code{\link{dpal}} \code{\link{spal}}
}

