% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotpal}
\alias{plotpal}
\title{Show color palette in a barplot}
\usage{
plotpal(colors, hex.labels = FALSE)
}
\arguments{
\item{colors}{vector of hex colors}

\item{hex.labels}{logical, should the colors hex label be shown on the plot}
}
\value{
\code{plotpal} returns a barchart showing the specified colors
}
\description{
\code{plotpal} allows you to easily visualize a color palette by plotting the
specified palette using a barchart
}
\examples{
plotpal(qpal(10))
plotpal(qpal(100))
plotpal(c("#00A0ED", "#F15A8F", "#4FA500"))
}
\seealso{
\code{\link{qpal}} \code{\link{dpal}} \code{\link{spal}}
}

