% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spal.R
\name{spal}
\alias{spal}
\title{Create sequential color palettes using the HCL color space}
\usage{
spal(n, hue.start = 235, hue.end = NULL, luminance = c(30, 85),
  chroma = c(90, 100), power = 1, alpha = 1, fixup = TRUE)
}
\arguments{
\item{n}{number of colors to create}

\item{hue.start}{starting hue value, should be in the interval [0, 360]}

\item{hue.end}{ending hue value, should be in the interval [0, 360].}

\item{luminance}{vector of length 2 giving the start and end bounds for the range of
luminance values}

\item{chroma}{vector of length 2 giving the start and end bounds for the range of chroma
values}

\item{power}{parameter to control how chroma and luminance increase/decrease}

\item{alpha}{transparency level defined on the interval [0, 1] where 0 =
transparent and 1 = opaque}

\item{fixup}{logical, should RGB value be corrected (see \code{\link{hcl}}}
}
\value{
\code{spal} returns a vector of hex colors
}
\description{
\code{spal} creates sequential color palettes using the HCL color space.
    Sequential color palettes are composed of a gradient of colors where the
    gradient can include a range of hues, luminance values, or chroma values.
    Sequential color palettes are suitable for encoding continuous variables
    where either low or high values are of importance, but not both.
}
\examples{
plotpal(spal(140))
plotpal(spal(140, luminance = c(20, 80)))
plotpal(spal(140, hue.start = 235, hue.end = 70))
plotpal(spal(140, hue.start = 235, hue.end = 70, luminance = c(20, 80)))

image(t(volcano)[ncol(volcano):1,], col = spal(100))
image(t(volcano)[ncol(volcano):1,], col = spal(100, hue.end = 70))

plotpal(spal(140, power = 0.8))
plotpal(spal(140, power = 1))
plotpal(spal(140, power = 1.3))
plotpal(spal(140, power = 1.5))

plotpal(spal(140, hue.start = 260, chroma = c(80, 0),
    luminance = c(30, 90), power = 1.5))

}
\seealso{
\code{\link{hcl}} \code{\link{qpal}} \code{\link{dpal}}
}
\author{
Michael Malick
}
