% Generated by roxygen2 (4.0.1): do not edit by hand
\name{colorwheel}
\alias{colorwheel}
\title{Visualize color palettes using a color wheel}
\usage{
colorwheel(chroma = 100, luminance = 60, labels = TRUE, alpha = 1,
  fixup = TRUE)
}
\arguments{
\item{chroma}{chroma value for palette}

\item{luminance}{luminance value for palette}

\item{labels}{logical, should degree labels be shown}

\item{alpha}{numeric vector of values in the range [0,1] for alpha
transparency channel (0 means transparent and 1 means opaque) #'}

\item{fixup}{a logical value which indicates whether the resulting RGB values
should be corrected to ensure that a real color results. if fixup is FALSE
RGB components lying outside the range [0,1] will result in an NA value}
}
\value{
\code{colorwheel} returns a pie chart showing the specified colors
}
\description{
\code{colorwheel} allows you to easily visualize a color palette by plotting
all hues for the specified specified chroma and luminance values using a
color wheel
}
\examples{
colorwheel()
colorwheel(chroma = 50, luminance = 60)
colorwheel(chroma = 100, luminance = 75)
colorwheel(chroma = 100, luminance = 25)
colorwheel(chroma = 100, luminance = 0)
colorwheel(chroma = 0, luminance = 50)

par(mfrow = c(4, 3), mar = c(2, 0, 1.5, 0))
for(i in seq(0, 100, 10))
   colorwheel(chroma = i, labels = TRUE, fixup = TRUE)

par(mfrow = c(3, 3), mar = c(2, 0, 1.5, 0))
for(i in seq(20, 100, 10))
    colorwheel(luminance = i, labels = TRUE)
}
\seealso{
\code{\link{qpal}} \code{\link{dpal}} \code{\link{spal}}
}

