% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dpal}
\alias{dpal}
\title{Create divergent color palettes using the HCL color space}
\usage{
dpal(n, hue = c(0, 235), chroma = c(0, 60), luminance = c(25, 100))
}
\arguments{
\item{n}{number of colors to create}

\item{hue}{vector of length 2 giving the first and second hues to use}

\item{chroma}{vector of length 2 giving the start and end bounds for the
range of chroma values}

\item{luminance}{vector of length 2 giving the start and end bounds for the
range of luminance values}
}
\value{
\code{dpal} returns a vector of hex colors
}
\description{
\code{dpal} creates divergent color palettes using the HCL color space.
Divergent color palettes consist of a gradient that goes from dark to light
to dark where both ends of the gradient are a separate hue.
Divergent color palettes  are suitable when low and high values are of
importance, but middle range values are less important.
}
\examples{
plotpal(dpal(200))
plotpal(dpal(200, hue = c(100, 240)))
plotpal(dpal(200, hue = c(40, 240), c = c(40, 100)))

mat <- matrix(c(rnorm(100), rnorm(100) + 1, rnorm(100) + 2, rnorm(100) + 3),
              ncol = 4)
image(z = mat, col = dpal(200), axes = FALSE)
box()
}
\seealso{
\code{\link{hcl}} \code{\link{qpal}} \code{\link{spal}}
}

