% Generated by roxygen2 (4.0.1): do not edit by hand
\name{spal}
\alias{spal}
\title{Create sequential color palettes using the HCL color space}
\usage{
spal(n, hue.start = 235, hue.end = NULL, luminance = c(30, 85),
  chroma = c(90, 100))
}
\arguments{
\item{n}{number of colors to create}

\item{hue.start}{starting hue value, should be in the interval [0, 360]}

\item{hue.end}{ending hue value, should be in the interval [0, 360].}

\item{chroma}{vector of length 2 giving the start and end bounds for the
range of chroma values}

\item{luminance}{vector of length 2 giving the start and end bounds for the
range of luminance values}
}
\value{
\code{spal} returns a vector of hex colors
}
\description{
\code{spal} creates sequential color palettes using the HCL color space.
Sequential color palettes are composed of a gradient of colors where the
gradient can include a range of hues, luminance values, or chroma values.
Sequential color palettes are suitable for encoding continuous variables
where either low or high values are of importance, but not both.
}
\examples{
plotpal(spal(140))
plotpal(spal(140, luminance = c(20, 80)))
plotpal(spal(140, hue.start = 235, hue.end = 70))
plotpal(spal(140, hue.start = 235, hue.end = 70, luminance = c(20, 80)))

image(t(volcano)[ncol(volcano):1,], col = spal(100))
image(t(volcano)[ncol(volcano):1,], col = spal(100, hue.end = 70))
}
\seealso{
\code{\link{hcl}} \code{\link{qpal}} \code{\link{dpal}}
}

